@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	
 	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.App" Version="2.1.5" />
+		
 	</ItemGroup>
 	
 	<ItemGroup>
@@ -13,7 +13,7 @@
 	
 	<PropertyGroup>
 		<OutputType>Exe</OutputType>
-		<TargetFramework>netcoreapp2.1</TargetFramework>
+		<TargetFramework>netcoreapp3.1</TargetFramework>
 		<StartupObject>Dvelop.Selfhosted.HostApplication.Program</StartupObject>
 		<RootNamespace>Dvelop.Selfhosted.HostApplication</RootNamespace>
 	</PropertyGroup>
